Package dependency system finished! Now cc-get recursively gets the dependencies for any package installed and also gets the dependencies of the dependencies.
Also, the package.ini format changes were incorporated. And the error message for an unknown command shows how to get help.
Oh, and the path is re-set after installing a package, so that doesn't need to be done by hand, although this can lead to things appearing in path multiple times until rebooting.